+++ src/routes/authRoutes.jsx
@@ -1,0 +1,10 @@
+import Login from "../pages/Login";
+
+const authRoutes = [
+  {
+    path: "/login",
+    element: <Login />,
+  },
+];
+
+export default authRoutes;
