+++ src/routes/homeRoutes.jsx
@@ -1,0 +1,10 @@
+import Home from "../pages/Home";
+
+const homeRoutes = [
+  {
+    path: "/",
+    element: <Home />,
+  },
+];
+
+export default homeRoutes;
